dev/cloud_functions/requirements.txt
feedparser==6.0.11
firebase-admin==6.5.0
google-cloud-translate==2.0.1